move feedback structures to commons package to use inside odahu-tools CLI
